pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-summary"
version = "0.1.0"
description = "A FastAPI service for text analysis and summarization"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.118.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "ruff>=0.13.2",
    "uvicorn>=0.37.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "isort",
]

[tool.hatch.build]
packages = ["llm_summary"]
